Extract base map functionalities to own component
